dtype=np.float32

+++ LFWDataset.py
@@ -35,7 +35,7 @@ class LFWDataset(torch.utils.data.Dataset):
         fn = os.path.join(dataLT, 'img0000.png')
         img = cv2.imread(fn)
         assert img.shape == LFWDataset.imageShape
-        X = np.empty((self.ndat,) + img.shape)
+        X = np.empty((self.ndat,) + img.shape, dtype=np.float32)
         for i in range(self.ndat):
             fn = os.path.join(dataLT, f'img{i:04d}.png')
             #print(fn)
